requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
asgiref==3.10.0
    # via
    #   django
    #   django-htmx
    #   django-stubs
asttokens==3.0.1
    # via stack-data
basedpyright==1.21.0
    # via -r requirements.in
certifi==2025.11.12
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via djlint
colorama==0.4.6
    # via djlint
coverage==7.12.0
    # via pytest-cov
cssbeautifier==1.15.4
    # via djlint
decorator==5.2.1
    # via ipython
django==5.1.4
    # via
    #   -r requirements.in
    #   django-extensions
    #   django-htmx
    #   django-picklefield
    #   django-q2
    #   django-redis
    #   django-stubs
    #   django-stubs-ext
django-extensions==3.2.3
    # via -r requirements.in
django-htmx==1.21.0
    # via -r requirements.in
django-picklefield==3.4.0
    # via django-q2
django-prometheus==2.3.1
    # via -r requirements.in
django-q2==1.9.0
    # via -r requirements.in
django-redis==6.0.0
    # via -r requirements.in
django-stubs==5.1.1
    # via -r requirements.in
django-stubs-ext==5.2.7
    # via django-stubs
djlint==1.36.2
    # via -r requirements.in
editorconfig==0.17.1
    # via
    #   cssbeautifier
    #   jsbeautifier
executing==2.2.1
    # via stack-data
factory-boy==3.3.3
    # via -r requirements.in
faker==38.2.0
    # via factory-boy
gunicorn==23.0.0
    # via -r requirements.in
idna==3.11
    # via requests
iniconfig==2.3.0
    # via pytest
ipython==8.29.0
    # via -r requirements.in
jedi==0.19.2
    # via ipython
jsbeautifier==1.15.4
    # via
    #   cssbeautifier
    #   djlint
json5==0.12.1
    # via djlint
matplotlib-inline==0.2.1
    # via ipython
nodejs-wheel-binaries==24.11.1
    # via basedpyright
packaging==25.0
    # via
    #   gunicorn
    #   pytest
    #   pytest-sugar
parso==0.8.5
    # via jedi
pathspec==0.12.1
    # via djlint
pexpect==4.9.0
    # via ipython
pluggy==1.6.0
    # via pytest
prometheus-client==0.23.1
    # via django-prometheus
prompt-toolkit==3.0.52
    # via ipython
psycopg2-binary==2.9.11
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pygments==2.19.2
    # via ipython
pytest==8.3.4
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-django
    #   pytest-sugar
pytest-cov==6.0.0
    # via -r requirements.in
pytest-django==4.9.0
    # via -r requirements.in
pytest-sugar==1.0.0
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
python-json-logger==3.2.1
    # via -r requirements.in
pyyaml==6.0.3
    # via djlint
redis==7.0.1
    # via django-redis
regex==2025.11.3
    # via djlint
requests==2.32.5
    # via -r requirements.in
ruff==0.8.4
    # via -r requirements.in
sentry-sdk==2.45.0
    # via -r requirements.in
six==1.17.0
    # via
    #   cssbeautifier
    #   jsbeautifier
sqlparse==0.5.3
    # via django
stack-data==0.6.3
    # via ipython
termcolor==3.2.0
    # via pytest-sugar
tqdm==4.67.1
    # via djlint
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
types-pyyaml==6.0.12.20250915
    # via django-stubs
typing-extensions==4.15.0
    # via
    #   django-stubs
    #   django-stubs-ext
tzdata==2025.2
    # via faker
ua-parser==1.0.1
    # via user-agents
ua-parser-builtins==0.18.0.post1
    # via ua-parser
urllib3==2.5.0
    # via
    #   requests
    #   sentry-sdk
user-agents==2.2.0
    # via -r requirements.in
wcwidth==0.2.14
    # via prompt-toolkit
whitenoise==6.8.2
    # via -r requirements.in
